apptemplate
Create a simple app template. Move Fancyclock and clock to misc
